Fix Issue #4: SFC Controller Board LED Footprint
This is embarassing and I'm not sure how it happened, but somehow
the LED footprint went missing. I have no idea what happened here
and apologize profusely to anyone affected.

The footprint was still there in the schema but became unlinked
from the board and raised a DRC issue. I have corrected this now
and will investigate to see if I can find out what happened.